+++ src/libs/database/repositories/SeasoningRepository.test.ts
@@ -4,7 +4,7 @@
  */
 
 import { describe, test, expect, beforeEach, vi } from "vitest";
-import type { IDatabaseConnection } from "@/libs/database/interfaces/IDatabaseConnection";
+import type { IDatabaseConnection } from "@/libs/database/interfaces";
 import type { SeasoningCreateInput } from "@/libs/database/interfaces/ISeasoningRepository";
 import { SeasoningRepository } from "./SeasoningRepository";
