+++ src/libs/database/repositories/SeasoningRepository.ts
@@ -15,7 +15,7 @@ import type {
   PaginatedResult,
 } from "@/libs/database/interfaces/common/types";
 import { Seasoning } from "@/libs/database/entities/Seasoning";
-import type { IDatabaseConnection } from "@/libs/database/interfaces/IDatabaseConnection";
+import type { IDatabaseConnection } from "@/libs/database/interfaces";
 
 /**
  * データベースから取得した生データの型定義
